core: split core/include/kernel/thread_private.h
Splits core/include/kernel/thread_private.h into one generic and one
architecture specific file.

Acked-by: Marouene Boubakri <marouene.boubakri@nxp.com>
Acked-by: Etienne Carriere <etienne.carriere@linaro.org>
